feat: 优惠券创建模板等系列方法

+++ coupon-template-serv/coupon-template-api/src/main/java/com/yejunyu/coupon/template/api/beans/request/PagedCouponTemplateInfo.java
@@ -22,5 +22,5 @@ public class PagedCouponTemplateInfo {
     // 第几页数据
     public int page;
     // 总数据量
-    public int total;
+    public long total;
 }
